Make counterbore holes deeper
I had made the upper bosses a bit longer, but hadn't adjusted the
counterbores. The result is that the screws wouldn't engage as
deeply. Change so the counterbore depth is relative to the bottom
of the boss instead of the case height.
Also tweaked some dimensions to make sure the screw head has enough plastic
to hold onto, but these may be below the abilities of the printer anyway.