Outside case radius was wrong.
I was using the boss radius without accounting for distance between it and the outside.
With the larger radius, I needed to decrease the chamfer width, otherwise it was excessive.
Also thickened walls to 2mm. The lip broke off in the previous print.
Removed blocks supporting bosses. There was a lot of warping near these.