Switch to a microcontroller with more SRAM
This is necessary to enable some effects like reverb and echo.
This part has a similar pinout, but not exactly the same, which required some layout
changes.
I also took this opportuntity to move the serial pins to a separate header with its
own ground, which makes connecting a serial debugger and SWD interface at the same
time easier.