Move I2C2_SDA from PA8 to PA10
There is conflicting information in the datahsheet. Table 13 says this
should be PA10, but Table 12 says PA8. I'm using the former, as it
has actual mappings. This would be fairly easy to rework with a blue
wire if I guessed wrong: jumpering the pins together and making PA10
be an input.