Use longer screws
The M2x6 screws are too short with the PCB in place. This requires upgrading to M2.3 screws.
Increased bore sizes. I had to make the top bosses tapered to accomodate the larger counterbore.
I tapered the bottom ones just to match.
Also increased the RELIEF to 0.8, as the parts do warp a bit and things stick during assembly.