Adjust enclosure a bit
- For lip/groove at seam, I had previously subtracted the relief from only
the bottom. Change so it is divided between the top and bottom so one isn't thinner/weaker than the other.
- I was previously subtracting the relief from the diameter of the button holes.
Instead, subtract twice the relief (or subtract the relief from the radius), so the relief ends up being the clearance all the way around the hole.
- Add a separate constant for the PCB inset from the inside wall, as it is quite a bit largerthan the relief.
- With all of these changes, crank the relief down to 0.5.
- Subtract the relief from the height of the bottom lip so it can fit snug.

Also cleaned up by changing all identifiers to lowercase. Rename some variables
to be clearer/more consistent.