Change how charge management is handled
The LDO (and microcontroller) are now enabled even if the power switch is off when the device is charging.
Added a bi-color LED to indicate charging status (driven by microcontroller)
Charge status now routed to the microcontroller, as well as power switch status.

(Note this would also allow data transfer to the microcontroller when plugged in when the power switch is off, which is probably desirable).